feat(tmux): restyle window list in status bar
This commit addresses a UX fail. Now the current window is the only item
with color. Activity is replaced by a "*" in blue. This makes it obvious
which window is active. This commit also disabled the message
notification on activity/bell since displaying a "*" is good enough and
less distracting.

This also updates config for Tmux 2.7.