feat(asdf): remove user .tool-versions
If a user specific .tool-versions is installed then ASDF will always use
whatever is declared in this file. This adds a hidden assumption for any
tool potentially managed with an ASDF plugin. For example, consider you
just kubectl and the version if irrelevant. If a ~/.tool-versions file
is present, then you _must_ use kubectl through ASDF _even if_ the
.tool-versions does not specify a kubectl version. The solution is
declare "system" for all tools managed through ASDF. This instructs ASDF
to act like a pass through to whatever is installed on the system.

However a better solution is to remove the ~/.tool-versions file and
only create a .tool-versions file for directories (i.e. code
repositories) that require it. This makes the behavior explicit, like,
"In this directory, tools are managed with ASDF".

The unexpected assumption creates more problems when dealing with tools
like node which may install their own global executables.

Generally speaking I think the best practice is to only create the
.tool-versions file explicitly when required. This leads to more
predictable outcomes.