feat(asdf): ignore missing plugins on direnv
This prevents a total failure when activating asdf and not all expected
versions are installed. This has bitten me a few times. It requires
rerunning things with a debugging flag to see the failure that "oh thing
version x is not installed". OK, just continue and allow me to debug
from there.